package refactoring for sample app
