+++ app/src/main/java/io/iamjosephmj/flingersample/ui/utils/NavigationSetup.kt
@@ -23,28 +23,22 @@
 *
 */
 
-package io.iamjosephmj.flinger.ui.utils
+package io.iamjosephmj.flingersample.ui.utils
 
-import androidx.compose.material.MaterialTheme
-import androidx.compose.material.lightColors
 import androidx.compose.runtime.Composable
-import androidx.compose.ui.graphics.Color
+import androidx.navigation.NavHostController
+import androidx.navigation.compose.NavHost
+import androidx.navigation.compose.composable
+import io.iamjosephmj.flingersample.ui.scroll.RenderScrollPage
+import io.iamjosephmj.flingersample.ui.settings.RenderSettingsPage
 
 @Composable
-fun LightTheme(
-    content: @Composable () -> Unit
-) {
-    MaterialTheme(
-        colors = LightColors,
-        content = content
-    )
+fun CreateNavHost(navController: NavHostController) {
+    NavHost(
+        navController = navController,
+        startDestination = "scrollPage",
+    ) {
+        composable("scrollPage") { RenderScrollPage(navController) }
+        composable("Settings") { RenderSettingsPage(navController) }
+    }
 }
-
-private val LightColors = lightColors(
-    primary = Color(0xFF009688),
-    primaryVariant = Color(0xFF009688),
-    secondary = Color(0xFF009688),
-    secondaryVariant = Color(0xFF018786),
-    background = Color.White,
-    surface = Color.White,
-    error = Color(0xFFB00020))
