+++ app/src/main/java/io/iamjosephmj/flingersample/ui/utils/Theme.kt
@@ -23,18 +23,28 @@
 *
 */
 
-package io.iamjosephmj.flinger.ui.utils
+package io.iamjosephmj.flingersample.ui.utils
 
-import android.text.TextUtils
+import androidx.compose.material.MaterialTheme
+import androidx.compose.material.lightColors
+import androidx.compose.runtime.Composable
+import androidx.compose.ui.graphics.Color
 
-fun String.toFloatNum(): Float {
-    return if (isNotEmpty() && length == 1 &&
-        TextUtils.equals(get(0).toString(), ".")
-    ) {
-        "0.".toFloat()
-    } else if (isEmpty()) {
-        "0".toFloat()
-    } else {
-        toFloat()
-    }
+@Composable
+fun LightTheme(
+    content: @Composable () -> Unit
+) {
+    MaterialTheme(
+        colors = LightColors,
+        content = content
+    )
 }
+
+private val LightColors = lightColors(
+    primary = Color(0xFF009688),
+    primaryVariant = Color(0xFF009688),
+    secondary = Color(0xFF009688),
+    secondaryVariant = Color(0xFF018786),
+    background = Color.White,
+    surface = Color.White,
+    error = Color(0xFFB00020))
